moved the work with cities to a separate file
